Fix signed/unsigned comparisons and migrations
Minor correctness and cleanup across modules:

- error_handler.cpp: Cast lines.size() to int for comparison to avoid signed/unsigned mismatch.
- menu_system.cpp: Same signed/unsigned fix for selection/index bounds checks.
- motor.cpp: Removed an unused temporary (fdaAmp) and simplified comments around frequency-dependent amplitude calculation; preserved the computed scaleFactor.
- settings.cpp: Reworked migration to populate defaults into a temporary newSettings, restore live state while reading legacy data, then assign newSettings to _data for V2 and V3 migrations to ensure new fields are initialized safely before copying legacy fields.
- ui.cpp: Simplified transition drawing by computing an xOffset from transition progress and applying it to title, lines, cursors, highlights and scrollbar; removed commented/duplicated transition logic.

+++ error_handler.cpp
@@ -96,7 +96,7 @@ void ErrorHandler::getLogLines(std::vector<String>& lines, int maxLines) {
     
     // Read lines from the beginning of the file
     // TODO: For large logs, implementing a tail reader would be more efficient
-    while (f.available() && lines.size() < maxLines) {
+    while (f.available() && (int)lines.size() < maxLines) {
         String line = f.readStringUntil('\n');
         line.trim();
         if (line.length() > 0) {
